omni_connect_app/requirements.txt

fastapi
uvicorn
python-dotenv
omnidimension
